电负荷预测/requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
openpyxl==3.0.10
mysql-connector-python==8.2.0
torch==2.0.1
pytorch-cuda==11.7
pytorch-mutex==1.0
torchaudio==2.0.2
torchvision==0.15.2
